Fixes filename and folder problem
previous files were just getting saved in the root directory of our
bucket. Solved this by adding today's date as a folder namer and
prepending out file name with that. The nest thing we did via
crypto.randomBytes is create a random string to save our file as on AWS.
The last thing that we did was promisify both s3.upload and
crypto.randomByres so both implement promises that are resolved or
rejected if they work or not.