Refactor bin script into module
Moves most of the code from bin/aws-upload that was responsible for the
s3 upload piece to its own file in the lib directory.
I saw the need for using this code in more than one place, so the
lib/se-upload module can now be used from more than place in our app.
The bin/aws-upload script still works and it simply implements the lin one